Make log system requires spdlog explicitly
To achieve that goal, we are using a macro
UT_USE_SPDLOG to ensure the user knows he
needs spdlog to use the log system.

Because the sandbox target is a sample
one, I did not define UT_USE_SPDLOG in the
xmake.lua file, but in the source code
directly.